pyproject.toml
[tool.poetry]
name = "e-commerce-api"  # Updated to remove spaces
version = "0.1.0"
description = ""
authors = ["Nhan Nguyen Van Khanh <nhan.nguyenvan@paradox.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.7"
amqp = "^5.2.0"
asgiref = "^3.8.1"
billiard = "^4.2.1"
black = "^24.10.0"
certifi = "^2024.8.30"
cffi = "^1.17.1"
charset-normalizer = "^3.4.0"
click = "^8.1.7"
click-didyoumean = "^0.3.1"
click-plugins = "^1.1.1"
click-repl = "^0.3.0"
cryptography = "^43.0.3"
django = "5.1.4"
django-cors-headers = "^4.6.0"
django-elasticsearch-dsl = "<8.0"
django-filter = "^24.3"
django-oauth-toolkit = "^3.0.1"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
elastic-transport = "^8.15.1"
elasticsearch = "<8"
elasticsearch-dsl = "<8"
factory-boy = "^3.3.1"
faker = "^30.8.1"
idna = "^3.10"
inflection = "^0.5.1"
jwcrypto = "^1.5.6"
kombu = "^5.4.2"
mypy = "^1.13.0"
mypy-extensions = "^1.0.0"
mysqlclient = "^2.2.5"
oauthlib = "^3.2.2"
packaging = "^24.1"
pathspec = "^0.12.1"
pillow = "^11.0.0"
platformdirs = "^4.3.6"
prompt-toolkit = "^3.0.48"
pycparser = "^2.22"
pyjwt = "^2.9.0"
python-dateutil = "^2.9.0.post0"
python-decouple = "^3.8"
pytz = "^2024.2"
pyyaml = "^6.0.2"
redis = "^5.2.0"
requests = "^2.32.3"
six = "^1.16.0"
sqlparse = "^0.5.1"
types-cffi = "^1.16.0.20240331"
types-pyopenssl = "^24.1.0.20240722"
types-redis = "^4.6.0.20241004"
types-setuptools = "^75.2.0.20241025"
typing-extensions = "^4.12.2"
tzdata = "^2024.2"
uritemplate = "^4.1.1"
urllib3 = "<2"
vine = "^5.1.0"
wcwidth = "^0.2.13"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"
icalendar = "^6.0.1"


[tool.poetry.group.dev.dependencies]
commitlint = "^1.3.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
